aegra/rag/pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "RAG application for SQL generation"
requires-python = ">=3.9"
dependencies = [
    "langchain>=1.0.0",
    "langchain-core>=0.2.0",
    "langchain-openai>=0.1.0",
    "chromadb>=0.4.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "openai>=1.0.0",
    "numpy>=1.20.0",
    "python-multipart"
]

[project.scripts]
rag = "main:main"
rag-run = "main:run_app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]